requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Django==5.1.2
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
docstring-to-markdown==0.15
idna==3.10
jedi==0.19.1
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.10
python-environ==0.4.54
requests==2.32.3
sqlparse==0.5.1
sslcommerz-lib==1.0
ujson==5.10.0
urllib3==2.2.3
